Squashed commit of the following:
commit d732087eba08ebebeffa0fd1b3cd1903c52be44e
Merge: eb74712 531234b
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Mon Apr 17 15:16:27 2023 +0200

    Merge pull request #55 from VallbonaProject/PV-40-Programar-Login

    Correcciones menores del Login

commit 531234bca6cf71ac0ab3d423a28e00e3c8811fbc
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Mon Apr 17 15:15:55 2023 +0200

    Correcciones menores del Login

commit eb747121469f491d8061ac57a1295d43537b2922
Merge: 3b757c7 010bb7f
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Fri Apr 14 19:10:30 2023 +0200

    Merge branch 'PV-14-USUARIO-Crear-Usuario-y-Modificar-Usuario-eliminar' into dev

commit 010bb7f6fe18d45fbb7db3a1718dc33ad24fb3c5
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Fri Apr 14 19:07:41 2023 +0200

    css puesto en archivo aparte y

commit 3b757c71adad2c2696274cbf216ae5ac3205ef77
Merge: 96467ca de0cf57
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Fri Apr 14 19:01:47 2023 +0200

    Merge pull request #54 from VallbonaProject/PV-40-Programar-Login

    Fallos de seguridad arreglados

commit de0cf57b66f5820ae4457280981dfb4f5fb53281
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Fri Apr 14 19:01:16 2023 +0200

    Fallos de seguridad arreglados

commit 96467cad08be53f4f4280f54cc84ee6828a80632
Merge: 0d013b1 cad1048
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Fri Apr 14 18:46:21 2023 +0200

    Merge pull request #53 from VallbonaProject/PV-40-Programar-Login

    Login añadido

commit cad10484b6cb5b9f0783afed05c2a94d44bd319d
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Fri Apr 14 18:45:53 2023 +0200

    Login añadido

commit 0d013b1755430ccf54f1d338f9691185727c71d0
Merge: ecd9e08 d5e2506
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Fri Apr 14 18:34:29 2023 +0200

    Merge branch 'PV-14-USUARIO-Crear-Usuario-y-Modificar-Usuario-eliminar' into dev

commit d5e250693403b03f824412d23957351e681294fa
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Fri Apr 14 17:55:51 2023 +0200

    Popup confirmar eliminacion

commit d9f3c0cbff942d66816fdc8329e296b331afdc36
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 19:25:54 2023 +0200

    creado modal para confirmar eliminacion (falta que elimine al darle a confirmar)

commit ecd9e0828061aa33f97bd4e35d46643b4986ae25
Merge: d8da163 de54f4c
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 19:23:07 2023 +0200

    Merge pull request #52 from VallbonaProject/PV-64-Dise-o-Detalles-Estada

    Detalles de contacto añadidos

commit de54f4c814eba5f8469604c30e86de13198896d7
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 19:22:39 2023 +0200

    Detalles de contacto añadidos

commit d8da163a93ca832805bb1c0eacab3e9276ca20ae
Merge: a4dd448 f40797b
Author: DavidLastreCorrales <113192337+DavidLastreCorrales@users.noreply.github.com>
Date:   Thu Apr 13 19:18:06 2023 +0200

    Merge pull request #51 from VallbonaProject/PV-61-Sidebar

    PV-61 Modificar vistas varias y sidebar

commit f40797b08c988e99907e96a77101f6364129b46d
Merge: fad745d a4dd448
Author: DavidLastreCorrales <113192337+DavidLastreCorrales@users.noreply.github.com>
Date:   Thu Apr 13 19:17:56 2023 +0200

    Merge branch 'dev' into PV-61-Sidebar

commit fad745dcb1a184dc5550cd15f8cd864769a40832
Author: DavidLastreCorrales <david_lastrecorrales@iescarlesvallbona.cat>
Date:   Thu Apr 13 19:15:19 2023 +0200

    PV-61 Modificar vistas varias y sidebar

commit a4dd448f9c024f0ab4c6b91bb43762bae3d38527
Merge: a67afcc 301f608
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 17:40:07 2023 +0200

    Merge pull request #50 from VallbonaProject/PV-63-Dise-o-Detalles-Contacto

    Detalles contacto y actualización de estilos

commit 301f608b570ad4ff6f96bc187925e095beaa640f
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 17:39:13 2023 +0200

    Detalles contacto i actualizacion de estilos

commit a67afcc1f8f7f8adb442ef243a6f20bc3c286333
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 17:08:19 2023 +0200

    el boton de cancelar del modal no funcionaba y ahora si

commit 248bfc645ca664668b2586177ee888b99d8ef9cc
Merge: 762dece 3d5977b
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 17:04:16 2023 +0200

    Merge branch 'PV-14-USUARIO-Crear-Usuario-y-Modificar-Usuario-eliminar' into dev

commit 3d5977b364d37394338ee13be8b4ffd138fcf7a5
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 16:42:44 2023 +0200

    Modal e icono para eliminar integrados en la tabla

commit 762dece868de2ea0f936ca7dc42cd724ad629cb5
Merge: 1565f98 643e8d2
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 16:17:18 2023 +0200

    Merge pull request #49 from VallbonaProject/PV-59-Dise-o-Detalles

    Actualizaciones menores de las paginas de Empresa

commit 889505bf90ea61b688eda82248a559543f46fba9
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 16:17:09 2023 +0200

    eliminado new.blade.php

commit 53c14a0d45662153f05a7023099fa8377afd3f66
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Thu Apr 13 16:08:28 2023 +0200

    Hecho el modal de usuarios correctamente y funcional

commit 1565f986da8618f86983236aafb9fe9f0b2b5849
Merge: 7a01000 19a7e1c
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Thu Apr 13 15:13:00 2023 +0200

    Merge pull request #48 from VallbonaProject/PV-59-Dise-o-Detalles

    Link per les valoracions del llistat Empresa

commit 59d705b07c51b834c327adb202e4d0cdb8a48dd7
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Wed Apr 12 20:56:36 2023 +0200

    Modal Users Hecho

commit 7a01000a0881fe135fd188fc2deb399df0cf5833
Merge: 92dfb9d 508c73e
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Wed Apr 12 20:40:46 2023 +0200

    Merge pull request #47 from VallbonaProject/PV-59-Dise-o-Detalles

    Nombre de tutor en los detalles de Empresa

commit 92dfb9d1d83f43f91837861f65b9ef674c2e5698
Merge: 1302ce9 090bbd5
Author: jmiguelcn <113700439+jmiguelcn@users.noreply.github.com>
Date:   Wed Apr 12 20:12:56 2023 +0200

    Merge pull request #46 from VallbonaProject/PV-59-Dise-o-Detalles

    Pagina de detalles de Empresa

commit 227e16ece96d149897e5e74ce67dafc7c85e58ac
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Wed Apr 12 17:33:06 2023 +0200

    Añadidos los selects y mejorada la vista principal

    - Se ha añadido selects para editar y crear usuarios.
    - También en la vista de la lista se ha cambiado los campos que iban por IDs por el texto del campo

commit 72df2c568dc13c4f638ffeca4b2578325e62fad9
Author: fernandezpau <pau_fernandezvillanueva@iescarlesvallbona.cat>
Date:   Wed Apr 12 16:58:29 2023 +0200

    Vista User Mejorada (faltan los selects de ciclo y rol)

+++ ProyectoFinal/app/Http/Controllers/UserController.php
@@ -8,24 +8,34 @@ use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 use App\Models\User;
+use App\Models\Cicle;
+use App\Models\Rol;
 
 class UserController extends BaseController
 {
-    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+  use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 
-    function list() 
-    { 
+  function list()
+  {
+    if (Auth::user()->rol_id == 5076) {
       $users = User::all();
+      $cicles = Cicle::all();
+      $rols = Rol::all();
 
-      return view('user.list', ['users' => $users]);
+      return view('user.list', ['users' => $users, 'cicles' => $cicles, 'rols' => $rols]);
+    } else {
+      return redirect('');
     }
+  }
 
-    function edit(Request $request, $id) 
-    { 
+  function edit(Request $request, $id)
+  {
+    if (Auth::user()->rol_id == 5076) {
       $user = User::find($id);
-      if ($request->isMethod('post')) {    
+      if ($request->isMethod('post')) {
         // recollim els camps del formulari en un objecte user
 
         //$user = new User;
@@ -41,14 +51,19 @@ class UserController extends BaseController
       }
       // si no venim de fer submit al formulari, hem de mostrar el formulari
 
-      $users = User::all();
+      $cicles = Cicle::all();
+      $rols = Rol::all();
 
-      return view('user.edit', ['users' => $users, 'user' => $user]);
+      return view('user.edit', ['cicles' => $cicles, 'rols' => $rols, 'user' => $user]);
+    } else {
+      return redirect('');
     }
+  }
 
-    function new(Request $request) 
-    { 
-      if ($request->isMethod('post')) {    
+  function new(Request $request)
+  {
+    if (Auth::user()->rol_id == 5076) {
+      if ($request->isMethod('post')) {
         // recollim els camps del formulari en un objecte user
 
         $user = new User;
@@ -64,16 +79,24 @@ class UserController extends BaseController
       }
       // si no venim de fer submit al formulari, hem de mostrar el formulari
 
-      $users = User::all();
+      $cicles = Cicle::all();
+      $rols = Rol::all();
 
-      return view('user.new', ['users' => $users]);
+      return view('user.new', ['cicles' => $cicles, 'rols' => $rols]);
+    } else {
+      return redirect('');
     }
+  }
 
-    function delete($id) 
-    { 
+  function delete($id)
+  {
+    if (Auth::user()->rol_id == 5076) {
       $user = User::find($id);
       $user->delete();
 
       return redirect()->route('user_list');
+    } else {
+      return redirect('');
     }
+  }
 }
